Move remote config to configs
