+++ core/lib/config/src/configs/en/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod remote;
