+++ core/lib/config/src/configs/mod.rs
@@ -9,6 +9,7 @@ pub use self::{
     da_client::{avail::AvailConfig, celestia::CelestiaConfig, eigen::EigenConfig, DAClientConfig},
     da_dispatcher::DADispatcherConfig,
     database::{DBConfig, PostgresConfig},
+    en::remote::RemoteENConfig,
     eth_sender::{EthConfig, GasAdjusterConfig},
     eth_watch::EthWatchConfig,
     experimental::{ExperimentalDBConfig, ExperimentalVmConfig, ExperimentalVmPlaygroundConfig},
@@ -48,6 +49,7 @@ pub mod contracts;
 pub mod da_client;
 pub mod da_dispatcher;
 pub mod database;
+mod en;
 pub mod eth_proof_manager;
 pub mod eth_sender;
 pub mod eth_watch;
