+++ core/lib/dal/src/external_node_config_dal.rs
@@ -1,4 +1,5 @@
-use zksync_db_connection::{connection::Connection, error::DalResult, instrument::InstrumentExt};
+use zksync_config::RemoteENConfig;
+use zksync_db_connection::{connection::Connection, error::DalResult, instrument::Instrumented};
 
 use crate::Core;
 
@@ -8,8 +9,12 @@ pub struct ExternalNodeConfigDal<'a, 'c> {
 }
 
 impl ExternalNodeConfigDal<'_, '_> {
-    pub async fn save_config(&mut self, value: serde_json::Value) -> DalResult<()> {
-        sqlx::query!(
+    pub async fn save_config(&mut self, value: &RemoteENConfig) -> DalResult<()> {
+        let instrumentation = Instrumented::new("save_en_config");
+
+        let value =
+            serde_json::to_value(value).map_err(|err| instrumentation.arg_error("value", err))?;
+        let query = sqlx::query!(
             r#"
             INSERT INTO
             en_remote_config (
@@ -23,16 +28,20 @@ impl ExternalNodeConfigDal<'_, '_> {
             updated_at = NOW()
             "#,
             value
-        )
-        .instrument("save_en_remote_config")
-        .execute(self.storage)
-        .await?;
+        );
+        instrumentation
+            .clone()
+            .with(query)
+            .execute(self.storage)
+            .await?;
 
         Ok(())
     }
 
-    pub async fn get_en_remote_config(&mut self) -> DalResult<Option<serde_json::Value>> {
-        let row = sqlx::query!(
+    pub async fn get_en_remote_config(&mut self) -> DalResult<Option<RemoteENConfig>> {
+        let instrumentation = Instrumented::new("get_en_remote_config");
+
+        let query = sqlx::query!(
             r#"
             SELECT
                 value
@@ -40,11 +49,17 @@ impl ExternalNodeConfigDal<'_, '_> {
                 en_remote_config
             LIMIT 1
             "#,
-        )
-        .instrument("get_last_notification_by_topics")
-        .fetch_optional(self.storage)
-        .await?;
+        );
 
-        Ok(row.map(|a| a.value))
+        let row = instrumentation
+            .clone()
+            .with(query)
+            .fetch_optional(self.storage)
+            .await?;
+        row.map(|a| {
+            serde_json::from_value(a.value)
+                .map_err(|err| instrumentation.arg_error("RemoteENConfig", err))
+        })
+        .transpose()
     }
 }
